Move reusable parts into separate package.
