+++ config/meta-cfg-to-toml.py
@@ -1,28 +1,13 @@
 #!/usr/bin/env python3
 from configparser import ConfigParser
-from toml_encoder import TomlArraySeparatorEncoderWithNewline
+from shared.call import call
+from shared.toml_encoder import TomlArraySeparatorEncoderWithNewline
 import argparse
 import collections
 import os
 import pathlib
-import subprocess
 import sys
 import toml
-
-
-def call(*args, capture_output=False, cwd=None):
-    """Call `args` as a subprocess.
-
-    If it fails exit the process.
-    """
-    result = subprocess.run(
-        args, capture_output=capture_output, text=True, cwd=cwd)
-    if result.returncode != 0:
-        print('ABORTING: Please fix the errors shown above.')
-        print('Proceed anyway (y/N)?', end=' ')
-        if input().lower() != 'y':
-            sys.exit(result.returncode)
-    return result
 
 
 parser = argparse.ArgumentParser(
